Improve performance of cmd_line_transformer
process_lines() doesn't need to split the input on newlines, save the
resulting lines in a table, and then create a new table with transformed
lines only to concatenate them again at the end. It can simply pass the
transform function to gsub() with a pattern matching everything except
new lines.